fix: Revert "feat: retry v8 (#916)"
This reverts commit e24406af3ccec0c56bef0b6e14ecc66f0d2fb77a.

# Conflicts:
#	packages/app/scripts/eas-build-pre-install.sh
