+++ packages/app/android/app/build.gradle
@@ -311,10 +311,10 @@ android {
     }
 
 
-    packagingOptions {
-        // Make sure libjsc.so does not packed in APK
-        exclude "**/libjsc.so"
-    }
+    // packagingOptions {
+    //     // Make sure libjsc.so does not packed in APK
+    //     exclude "**/libjsc.so"
+    // }
 }
