+++ packages/app/android/app/src/main/java/so/onekey/app/wallet/MainApplication.java
@@ -31,7 +31,7 @@ import java.lang.reflect.InvocationTargetException;
 import java.util.List;
 import android.webkit.WebView;
 
-import io.csie.kudo.reactnative.v8.executor.V8ExecutorFactory;
+// import io.csie.kudo.reactnative.v8.executor.V8ExecutorFactory;
 
 public class MainApplication extends Application implements ReactApplication , ViewModelStoreOwner {
   private final ViewModelStore mViewModelStore = new ViewModelStore();
@@ -61,14 +61,14 @@ public class MainApplication extends Application implements ReactApplication , V
       return "__generated__/AppEntry.js";
     }
 
-    @Override
-    protected JavaScriptExecutorFactory getJavaScriptExecutorFactory() {
-      return new V8ExecutorFactory(
-          getApplicationContext(),
-          getPackageName(),
-          AndroidInfoHelpers.getFriendlyDeviceName(),
-          getUseDeveloperSupport());
-    }
+    // @Override
+    // protected JavaScriptExecutorFactory getJavaScriptExecutorFactory() {
+    //   return new V8ExecutorFactory(
+    //       getApplicationContext(),
+    //       getPackageName(),
+    //       AndroidInfoHelpers.getFriendlyDeviceName(),
+    //       getUseDeveloperSupport());
+    // }
 
   });
