+++ packages/app/package.json
@@ -39,6 +39,7 @@
     "expo-location": "^14.2.2",
     "expo-media-library": "^14.1.0",
     "expo-splash-screen": "^0.15.1",
+    "jsc-android": "npm:@onekeyfe/jsc-android",
     "lottie-ios": "3.2.3",
     "lottie-react": "^2.2.1",
     "lottie-react-native": "5.0.1",
@@ -61,20 +62,18 @@
     "react-native-keyboard-manager": "^6.5.4-4",
     "react-native-pager-view": "5.4.15",
     "react-native-randombytes": "^3.6.1",
-    "react-native-reanimated": "npm:@onekeyfe/react-native-reanimated",
+    "react-native-reanimated": "^2.8.0",
     "react-native-restart": "0.0.24",
     "react-native-safe-area-context": "4.2.4",
     "react-native-screens": "3.11.1",
     "react-native-svg": "^12.3.0",
     "react-native-toast-message": "2.1.2",
-    "react-native-v8": "^1.4.1",
     "react-native-vision-camera": "^2.13.3",
     "react-native-web": "0.17.7",
     "react-native-webview": "11.18.1",
     "readable-stream": "^3.6.0",
     "realm": "10.20.0-beta.2",
     "text-encoding": "^0.7.0",
-    "v8-android-jit": "^10.100.0",
     "vision-camera-code-scanner": "^0.2.0"
   },
   "devDependencies": {
