+++ packages/app/react-native.config.js
@@ -1,16 +1,16 @@
 module.exports = {
-  dependencies: {
-    // disable frameprocessor and vision-camera-code-scanner
-    // until reanimated v3 offcially supports v8
-    'vision-camera-code-scanner': {
-      platforms: {
-        android: null, // disable Android platform, other platforms will still autolink if provided
-      },
-    },
-    'react-native-vision-camera': {
-      platforms: {
-        android: null, // disable Android platform, other platforms will still autolink if provided
-      },
-    },
-  },
+  // dependencies: {
+  //   // disable frameprocessor and vision-camera-code-scanner
+  //   // until reanimated v3 offcially supports v8
+  //   'vision-camera-code-scanner': {
+  //     platforms: {
+  //       android: null, // disable Android platform, other platforms will still autolink if provided
+  //     },
+  //   },
+  //   'react-native-vision-camera': {
+  //     platforms: {
+  //       android: null, // disable Android platform, other platforms will still autolink if provided
+  //     },
+  //   },
+  // },
 };
