+++ packages/kit/src/views/ScanQrcode/ScanCamera.tsx
@@ -1,7 +1,12 @@
 import React, { FC } from 'react';
 
-import { BarCodeScanner } from 'expo-barcode-scanner';
-import { Camera } from 'expo-camera';
+import { runOnJS } from 'react-native-reanimated';
+import {
+  Camera,
+  useCameraDevices,
+  useFrameProcessor,
+} from 'react-native-vision-camera';
+import { BarcodeFormat, scanBarcodes } from 'vision-camera-code-scanner';
 
 import { ScanCameraProps } from './types';
 
@@ -10,19 +15,30 @@ const ScanCamera: FC<ScanCameraProps> = ({
   isActive,
   children,
   onQrcodeScanned,
-}) =>
-  isActive ? (
-    <Camera
-      style={style}
-      onBarCodeScanned={({ data }) => onQrcodeScanned(data)}
-      barCodeScannerSettings={{
-        // eslint-disable-next-line @typescript-eslint/no-unsafe-member-access
-        barCodeTypes: [BarCodeScanner.Constants.BarCodeType.qr],
-      }}
-    >
+}) => {
+  const devices = useCameraDevices();
+  const device = devices.back;
+  const frameProcessor = useFrameProcessor((frame) => {
+    'worklet';
+
+    const detectedBarcodes = scanBarcodes(frame, [BarcodeFormat.QR_CODE], {
+      checkInverted: true,
+    });
+    runOnJS(onQrcodeScanned)(detectedBarcodes[0]?.rawValue);
+  }, []);
+  return device ? (
+    <>
+      <Camera
+        style={style}
+        device={device}
+        isActive={isActive}
+        frameProcessor={frameProcessor}
+        frameProcessorFps={5}
+      />
       {children}
-    </Camera>
+    </>
   ) : null;
+};
 ScanCamera.displayName = 'ScanCamera';
 
 export default ScanCamera;
